remove file only if buffer was provided

+++ controller/sending/sendFiles.ts
@@ -135,7 +135,8 @@ export async function sendFileHttpMethod(
     console.error(e);
   } finally {
     try {
-      await fsPromises.unlink(`${fileName}`);
+      if (buffer)
+        await fsPromises.unlink(`${fileName}`);
     } catch (e) {
       console.error(e);
     }
